adding more debug messages
the default epl template handler provides a few debug messages
that help figuring out what's going on. This patch adds that
same debug messages to the haml handler.

One important thing to notice is that a missing template is
not an error anymore, but a debug message. This is the same as
the default handler, though it might be interesting to make
it at least an 'info' level message.